feat(upload): factor cmd logic out to separate file
